refactor: split quadlet module under the 500-line limit
Extract the unit builders, render helpers and warning collection out of
quadlet/mod.rs into unit.rs, render.rs and warnings.rs. Pure mechanical
split; behavior unchanged, all quadlet tests pass.
