+++ internal/quadlet/mod.rs
@@ -11,12 +11,12 @@
 //! every field that is set but has no Quadlet equivalent yet, so generated
 //! units never quietly drop configuration.
 
-use std::collections::BTreeMap;
+mod render;
+mod unit;
+mod warnings;
 
-use crate::compose::types::{
-	Command, ComposeFile, PortMapping, RestartPolicy, Service, VolumeMount,
-};
-use crate::ports::parse_ports;
+use crate::compose::types::ComposeFile;
+use unit::{container_unit, network_unit, volume_unit};
 
 /// A single generated unit file: its name and full contents.
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -64,334 +64,6 @@ pub fn generate(file: &ComposeFile, project: &str) -> QuadletOutput {
 	}
 
 	out
-}
-
-fn network_unit(name: &str, project: &str, _has_config: bool) -> QuadletUnit {
-	let value = sanitize_value(&format!("{project}_{name}"));
-	let contents =
-		format!("[Network]\nNetworkName={value}\n\n[Install]\nWantedBy=default.target\n");
-	QuadletUnit {
-		filename: format!("{}.network", safe_unit_stem(name)),
-		contents,
-	}
-}
-
-fn volume_unit(name: &str, project: &str, _has_config: bool) -> QuadletUnit {
-	let value = sanitize_value(&format!("{project}_{name}"));
-	let contents =
-		format!("[Volume]\nVolumeName={value}\n\n[Install]\nWantedBy=default.target\n");
-	QuadletUnit {
-		filename: format!("{}.volume", safe_unit_stem(name)),
-		contents,
-	}
-}
-
-fn container_unit(
-	name: &str,
-	service: &Service,
-	declared_volumes: &[&str],
-	warnings: &mut Vec<String>,
-) -> QuadletUnit {
-	let mut unit = Section::new("Unit");
-	unit.add("Description", format!("{name} (podup)"));
-	for dep in service.depends_on.service_names() {
-		unit.add("After", format!("{dep}.service"));
-		if service.depends_on.required_for(&dep) {
-			unit.add("Requires", format!("{dep}.service"));
-		} else {
-			unit.add("Wants", format!("{dep}.service"));
-		}
-	}
-
-	let mut container = Section::new("Container");
-	container.add("ContainerName", name.to_string());
-	if let Some(image) = &service.image {
-		container.add("Image", image.clone());
-	}
-	if let Some(hostname) = &service.hostname {
-		container.add("HostName", hostname.clone());
-	}
-	if let Some(user) = &service.user {
-		container.add("User", user.clone());
-	}
-	if let Some(wd) = &service.working_dir {
-		container.add("WorkingDir", wd.clone());
-	}
-	if service.read_only == Some(true) {
-		container.add("ReadOnly", "true".to_string());
-	}
-	if service.init == Some(true) {
-		container.add("RunInit", "true".to_string());
-	}
-
-	match parse_ports(&service.ports) {
-		Ok(ports) => {
-			for p in ports {
-				container.add("PublishPort", render_publish_port(&p));
-			}
-		}
-		Err(_) => {
-			// Fall back to the raw short forms so nothing is dropped.
-			for port in &service.ports {
-				if let PortMapping::Short(s) = port {
-					container.add("PublishPort", s.clone());
-				}
-			}
-		}
-	}
-
-	for (key, val) in sorted_pairs(service.environment.to_map()) {
-		match val {
-			Some(v) => container.add("Environment", format!("{key}={v}")),
-			None => container.add("Environment", key),
-		}
-	}
-
-	for vol in &service.volumes {
-		container.add("Volume", render_volume(vol, declared_volumes));
-	}
-	for net in service.networks.names() {
-		container.add("Network", format!("{net}.network"));
-	}
-	for (key, val) in sorted_label_pairs(service.labels.to_map()) {
-		container.add("Label", format!("{key}={val}"));
-	}
-	for cap in &service.cap_add {
-		container.add("AddCapability", cap.clone());
-	}
-	for cap in &service.cap_drop {
-		container.add("DropCapability", cap.clone());
-	}
-	if let Some(entrypoint) = &service.entrypoint {
-		container.add("Entrypoint", render_command(entrypoint));
-	}
-	if let Some(command) = &service.command {
-		container.add("Exec", render_command(command));
-	}
-
-	let mut svc = Section::new("Service");
-	if let Some(restart) = &service.restart {
-		svc.add("Restart", render_restart(restart));
-	}
-
-	collect_warnings(name, service, warnings);
-
-	let mut contents = String::new();
-	contents.push_str(&unit.render());
-	contents.push('\n');
-	contents.push_str(&container.render());
-	if !svc.is_empty() {
-		contents.push('\n');
-		contents.push_str(&svc.render());
-	}
-	contents.push_str("\n[Install]\nWantedBy=default.target\n");
-
-	QuadletUnit {
-		filename: format!("{}.container", safe_unit_stem(name)),
-		contents,
-	}
-}
-
-/// Warn for fields that are set but have no Quadlet mapping, so the operator
-/// knows the generated unit is incomplete rather than discovering it at run
-/// time.
-fn collect_warnings(name: &str, service: &Service, warnings: &mut Vec<String>) {
-	let mut warn = |field: &str, detail: &str| {
-		warnings.push(format!("{name}: {field} {detail}"));
-	};
-	if service.build.is_some() {
-		warn(
-			"build",
-			"has no Quadlet equivalent; build the image first and set `image`",
-		);
-	}
-	let replicas = service
-		.scale
-		.or(service.deploy.as_ref().and_then(|d| d.replicas));
-	if replicas.is_some_and(|r| r > 1) {
-		warn(
-			"scale/replicas",
-			"is ignored; Quadlet emits a single container per service",
-		);
-	}
-	if service.healthcheck.is_some() {
-		warn("healthcheck", "is not yet mapped to HealthCmd directives");
-	}
-	if !service.secrets.is_empty() {
-		warn(
-			"secrets",
-			"are not yet mapped to Quadlet Secret= directives",
-		);
-	}
-	if !service.configs.is_empty() {
-		warn("configs", "have no Quadlet equivalent and are skipped");
-	}
-	if !service.volumes_from.is_empty() {
-		warn("volumes_from", "has no Quadlet equivalent and is skipped");
-	}
-	if service.network_mode.is_some() {
-		warn("network_mode", "is not mapped; use networks instead");
-	}
-	if !service.profiles.is_empty() {
-		warn("profiles", "have no Quadlet equivalent and are ignored");
-	}
-	if service.privileged == Some(true) {
-		warn(
-			"privileged",
-			"is not mapped; add PodmanArgs manually if required",
-		);
-	}
-}
-
-fn render_publish_port(p: &crate::ports::ParsedPort) -> String {
-	let mut s = String::new();
-	if !p.host_ip.is_empty() {
-		s.push_str(&p.host_ip);
-		s.push(':');
-	}
-	// A host port of None/0 means "let Podman pick"; omit it so the published
-	// side is empty (PublishPort=<container>) and Podman assigns a port.
-	if let Some(host) = p.host_port.filter(|n| *n != 0) {
-		s.push_str(&host.to_string());
-		s.push(':');
-	}
-	s.push_str(&p.container_port.to_string());
-	if p.protocol != "tcp" {
-		s.push('/');
-		s.push_str(&p.protocol);
-	}
-	s
-}
-
-fn render_volume(vol: &VolumeMount, declared_volumes: &[&str]) -> String {
-	match vol {
-		VolumeMount::Short(s) => {
-			let parts: Vec<&str> = s.splitn(3, ':').collect();
-			if parts.len() >= 2 && declared_volumes.contains(&parts[0]) {
-				let mut out = format!("{}.volume:{}", parts[0], parts[1]);
-				if let Some(opts) = parts.get(2) {
-					out.push(':');
-					out.push_str(opts);
-				}
-				out
-			} else {
-				s.clone()
-			}
-		}
-		VolumeMount::Long {
-			source,
-			target,
-			read_only,
-			..
-		} => {
-			let src = source.clone().unwrap_or_default();
-			let src = if declared_volumes.contains(&src.as_str()) {
-				format!("{src}.volume")
-			} else {
-				src
-			};
-			let mut out = if src.is_empty() {
-				target.clone()
-			} else {
-				format!("{src}:{target}")
-			};
-			if *read_only == Some(true) {
-				out.push_str(":ro");
-			}
-			out
-		}
-	}
-}
-
-fn render_command(command: &Command) -> String {
-	match command {
-		Command::Shell(s) => s.clone(),
-		Command::Exec(parts) => parts.join(" "),
-	}
-}
-
-fn render_restart(restart: &RestartPolicy) -> String {
-	match restart {
-		RestartPolicy::No => "no".to_string(),
-		RestartPolicy::Always => "always".to_string(),
-		RestartPolicy::UnlessStopped => "always".to_string(),
-		RestartPolicy::OnFailure { .. } => "on-failure".to_string(),
-	}
-}
-
-fn sorted_pairs(
-	map: std::collections::HashMap<String, Option<String>>,
-) -> Vec<(String, Option<String>)> {
-	let sorted: BTreeMap<_, _> = map.into_iter().collect();
-	sorted.into_iter().collect()
-}
-
-fn sorted_label_pairs(map: std::collections::HashMap<String, String>) -> Vec<(String, String)> {
-	let sorted: BTreeMap<_, _> = map.into_iter().collect();
-	sorted.into_iter().collect()
-}
-
-/// Strip ASCII control characters from a value before it is written into a
-/// unit file. systemd unit entries are single-line `Key=Value` pairs; an
-/// embedded newline from a hostile compose field would otherwise inject
-/// arbitrary unit directives (e.g. a `[Service]` `ExecStartPre=`). Compose
-/// input is untrusted, so every dynamic value is sanitized at the boundary.
-fn sanitize_value(value: &str) -> String {
-	value.chars().filter(|c| !c.is_control()).collect()
-}
-
-/// Reduce a compose key to a safe single path-component stem for a unit file
-/// name. Keeps ASCII alphanumerics and `-`/`_`/`.`; replaces anything else
-/// (path separators, control characters) with `_`, and guarantees the result
-/// is non-empty and does not start with a dot, so it can never escape the
-/// output directory or resolve to `.`/`..`.
-fn safe_unit_stem(name: &str) -> String {
-	let mut stem: String = name
-		.chars()
-		.map(|c| {
-			if c.is_ascii_alphanumeric() || matches!(c, '-' | '_' | '.') {
-				c
-			} else {
-				'_'
-			}
-		})
-		.collect();
-	if stem.is_empty() || stem.starts_with('.') {
-		stem.insert(0, '_');
-	}
-	stem
-}
-
-/// A single `[Section]` accumulating `Key=Value` lines in insertion order.
-struct Section {
-	name: &'static str,
-	lines: Vec<String>,
-}
-
-impl Section {
-	fn new(name: &'static str) -> Self {
-		Section {
-			name,
-			lines: Vec::new(),
-		}
-	}
-
-	fn add(&mut self, key: &str, value: String) {
-		self.lines.push(format!("{key}={}", sanitize_value(&value)));
-	}
-
-	fn is_empty(&self) -> bool {
-		self.lines.is_empty()
-	}
-
-	fn render(&self) -> String {
-		let mut s = format!("[{}]\n", self.name);
-		for line in &self.lines {
-			s.push_str(line);
-			s.push('\n');
-		}
-		s
-	}
 }
 
 #[cfg(test)]
@@ -657,23 +329,6 @@ services:
 			!c.lines().any(|l| l.starts_with("ExecStartPre")),
 			"a newline in a value must not inject a directive line:\n{c}"
 		);
-	}
-
-	#[test]
-	fn safe_unit_stem_neutralizes_traversal_and_control_chars() {
-		assert_eq!(safe_unit_stem("web"), "web");
-		assert_eq!(safe_unit_stem("db-data_1.x"), "db-data_1.x");
-		assert_eq!(safe_unit_stem("../../etc/passwd"), "_.._.._etc_passwd");
-		assert_eq!(safe_unit_stem("/abs"), "_abs");
-		assert_eq!(safe_unit_stem(".hidden"), "_.hidden");
-		assert_eq!(safe_unit_stem(""), "_");
-		assert!(!safe_unit_stem("a\nb").contains('\n'));
-	}
-
-	#[test]
-	fn sanitize_value_strips_control_characters() {
-		assert_eq!(sanitize_value("plain"), "plain");
-		assert_eq!(sanitize_value("a\nb\tc\r"), "abc");
 	}
 
 	#[test]
